General improvements on metrics and loss calculations. ResNet34 snapshot.
This commit makes the visibility of keypoints easy to include with
predefined helper functions that will give respective losses with quick
to access methods. These need to called in a functional way to be able
to have tensorflow compute the gradient accordingly. Additionally
renames have been done for an upcoming rework of CPMs and a snapshot of
the uploaded run of ResNet34 is included.

Signed-off-by: Mickey Vänskä <mvaenskae@gmail.com>